Move pending author navigation effect
