+++ hooks/usePushNotifications.ts
@@ -6,13 +6,12 @@ import Constants from 'expo-constants';
 import { Platform } from 'react-native';
 import { navigationRef, navigateToAuthorDetail } from '../navigation/navigationRef';
 import { PostInfo } from '../types';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 import {
   lastNotificationState,
   pushTokenState,
   pendingAuthorState,
 } from '../states/pushNotificationState';
-import { navigationReadyState } from '../states/navigationState';
 
 
 // https://docs.expo.dev/versions/latest/sdk/notifications/
@@ -20,7 +19,6 @@ export function usePushNotifications() {
   const [expoPushToken, setExpoPushToken] = useRecoilState(pushTokenState);
   const setNotification = useSetRecoilState(lastNotificationState);
   const [pendingAuthor, setPendingAuthor] = useRecoilState(pendingAuthorState);
-  const navigationReady = useRecoilValue(navigationReadyState);
 
   const notificationListener = useRef<Notifications.EventSubscription>();
   const responseListener = useRef<Notifications.EventSubscription>();
@@ -67,13 +65,6 @@ export function usePushNotifications() {
     };
   }, [setExpoPushToken, setNotification, setPendingAuthor]);
 
-  // 在導航就緒且存在待處理作者時進行跳轉
-  useEffect(() => {
-    if (navigationReady && pendingAuthor && navigationRef.isReady()) {
-      navigateToAuthorDetail(pendingAuthor);
-      setPendingAuthor(null);
-    }
-  }, [navigationReady, pendingAuthor, setPendingAuthor]);
 
   return {
     expoPushToken,
